+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - DB2 테스트 클래스 추가
 - DbManager JOIN 관련 테스트
 - DbManager GroupBy 관련 테스트
+- DbManager Sub Query 관련 테스트
 
 ### - 2024-10-25
 - yml 리소스 limit 추가 및 업데이트